[i18n] Move LongCodeToString to .cpp - shuts up mistaken clang warning about unused function
